+++ src/views/client-manage/index.vue
@@ -69,6 +69,7 @@
         <template slot-scope="scope">
           <el-button
             size="mini"
+            type="primary"
             @click="handleEdit(scope.row)"
           >编辑</el-button>
           <el-button
@@ -182,16 +183,25 @@ export default {
 
     // 删除
     handleDelete(row) {
-      deleteClient(row.id).then(
-        res => {
-          this.$message({
-            message: '删除成功',
-            type: 'success'
-          })
-          this.pageNum = 1
-          this.loadData()
-        }
-      )
+      this.$confirm('确定删除该客户端吗', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        showClose: false,
+        type: 'warning'
+      }).then(() => {
+        deleteClient(row.id).then(
+          res => {
+            this.$message({
+              message: '删除成功',
+              type: 'success'
+            })
+            this.pageNum = 1
+            this.loadData()
+          }
+        )
+      }).catch(() => {
+        // TODO
+      })
     },
 
     // 弹框关闭
